Image upload admin (#87)
* add upload ifor admin

* Development (#86)

* Development (#81)

* add prisma and mysql

* 3g memory

* fix config docker

* move to public

* add admin

* Config docker setup & DB connection with Prisma ORM (#66)

* Config docker setup & DB connection with Prisma ORM

* Add .env example (#67)

* Fix hot reload for dev (#68)

* Config docker setup & DB connection with Prisma ORM

* Add .env example (#67)

* Fix hot reload and link

* Fix prisma version and add generate script in Dockerfile for production (#69)

* Add User with role and role checking, auth and validation (#70)

* Hot fix pagecontainer (#71)

* Add enpoints and methods for Article and etc (#72)

* Add optimization builders, policies, concern (#73)

* Add user update (#74)

* Add optimization builders, policies, concern

* Add change role for user, add update user

* Create structure for studio (#75)

* Add api builder for article (#76)

* save before pull

* Development (#77)

* Add optimization builders, policies, concern (#73)

* Add user update (#74)

* Add optimization builders, policies, concern

* Add change role for user, add update user

* Create structure for studio (#75)

* Add api builder for article (#76)

---------


* Hot fix (#78)

* save before pull

* Articles admin (#80)

* Development (#79)

* Add optimization builders, policies, concern (#73)

* Add user update (#74)

* Add optimization builders, policies, concern

* Add change role for user, add update user

* Create structure for studio (#75)

* Add api builder for article (#76)

* Hot fix (#78)

---------


* add crud article by admin

---------


* change asynchrone react types

---------


* Add script to pass reset (#83)

* Add sliders to article (#84)

* Fix article update (#85)

* Add sliders to article

* Add fix

---------


* add galery and fix bugs article admin

---------

+++ src/api/services/media.service.ts
@@ -49,10 +49,20 @@ export const mediaService = {
     },
 
     async delete(user: any, id: number) {
-        if (!user) throw new ApiError(401, "Unauthorized");
-        if (user.role !== "ADMIN")
+    if (!user) throw new ApiError(401, "Unauthorized");
+    if (user.role !== "ADMIN")
         throw new ApiError(403, "Only ADMIN can delete media");
 
-        return mediaRepository.delete(id);
+    const media = await mediaRepository.findById(id);
+    if (!media) throw new ApiError(404, "Media not found");
+
+    const filePath = path.join(process.cwd(), "public", media.url);
+
+    try {
+        fs.unlinkSync(filePath);
+    } catch {
+    }
+
+    return mediaRepository.delete(id);
     },
 };
